requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
deprecation==2.1.0
Django==4.1.6
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
img2pdf==0.4.4
kavenegar==1.1.2
lxml==4.9.2
packaging==23.0
pikepdf==6.2.9
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
